Update code so that the stream event source is service independent
This way both, DynamoDB and Kinesis streams can be used with the help of the
"stream" event rather than two different event types ("dynamodb" and "kinesis").
