+++ docs/02-providers/aws/04-resource-names-reference.md
@@ -34,4 +34,4 @@ We're also using the term `normalizedName` or similar terms in this guide. This 
 |ApiGateway::Deployment | ApiGatewayDeployment{randomNumber}                      | ApiGatewayDeployment12356789  |
 |ApiGateway::ApiKey     | ApiGatewayApiKey{SequentialID}                          | ApiGatewayApiKey1             |
 |SNS::Topic             | SNSTopic{normalizedTopicName}                           | SNSTopicSometopic             |
-|AWS::Lambda::EventSourceMapping | <ul><li>**DynamoDB**: {normalizedFunctionName}EventSourceMappingDynamoDb{tableName} </li></ul> | <ul><li>**DynamoDB**: HelloLambdaEventSourceMappingDynamoDbUsers </li></ul> |
+|AWS::Lambda::EventSourceMapping | <ul><li>**DynamoDB**: {normalizedFunctionName}EventSourceMappingDynamodb{tableName} </li><li>**Kinesis**: {normalizedFunctionName}EventSourceMappingKinesis{streamName} </li></ul> | <ul><li>**DynamoDB**: HelloLambdaEventSourceMappingDynamodbUsers </li><li>**Kinesis**: HelloLambdaEventSourceMappingKinesisMystream </li></ul> |
