+++ docs/02-providers/aws/events/README.md
@@ -10,5 +10,4 @@ layout: Doc
 * [S3](./02-s3.md)
 * [Schedule](./03-schedule.md)
 * [SNS](./04-sns.md)
-* [Kinesis Streams](./05-kinesis-streams.md)
-* [DynamoDB Streams](./06-dynamodb-streams.md)
+* [DynamoDB / Kinesis Streams](./05-streams.md)
