+++ lib/plugins/Plugins.json
@@ -18,7 +18,7 @@
     "./aws/deploy/compile/events/s3/index.js",
     "./aws/deploy/compile/events/apiGateway/index.js",
     "./aws/deploy/compile/events/sns/index.js",
-    "./aws/deploy/compile/events/dynamodb/index.js",
+    "./aws/deploy/compile/events/stream/index.js",
     "./aws/deployFunction/index.js"
   ]
 }
